rename file fraction_scan.py
